+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Add support for Python 3.14 ([#2022](../../pull/2022), [#2026](../../pull/2026))
 - In Jupyter, set edge value of tilesource to transparent ([#2019](../../pull/2019))
 
+### Bug Fixes
+
+- When reading raw files via the PIL source, fix handling uint16 data ([#2029](../../pull/2029))
+
 ## 1.33.5
 
 ### Improvements